Change tsig module to work without std. (#152)
This commit modifies the tsig module to work without std. Unfortunately,
this means that a number of functions need to be changed to get the current
time passed in as SystemTime is only available in std. However, I think
allowing the user to provide the current time explicitly is better.

This is a breaking change.