Remove Deref and DerefMut from types that shouldn’t have them. (#205)
This PR removes the Deref and DerefMut impls from types that aren’t smart
pointers and thus shouldn’t really have them according to the advise given
in the documentation.

Where appropriate, it adds len and is_empty methods. Additional methods
likely should be added later as well.

In particular:

   o  Removed impls of Deref and (if present) DerefMut for the following
      types: CharStr, CharStrBuilder, DnameBuilder, Label, A, Aaaa, all
      record types that wrap a domain name, Nsec3Salt, OwnerHash,
      SearchList, ServerList. Where necessary, len and is_empty were added.

   o  IntoIterator for Label has a changed Self::Item.

This is a breaking change.