Move `flatten_into` for flattening domain names into a trait. (#216)
This PR moves the flatten_into method present on various types that either
are or contain domain names into a FlattenInto trait. It also removes the
method from all types that aren’t or don’t contain domain names and only had
it to be able to make the big record data type enums.