Fix displaying of chained domain names. (#253)
This PR fixes the Display implementation of Chain<_,_> so it considers all
possible combinations of empty, non-empty, and root names. It now behaves
similarly to the Display implementation of Dname in that it only adds a final
dot to an absolute name if it is the root label only. Similarly to Dname, it
also adds Chain::fmt_with_dot which returns a temporary type that formats
absolute names with a final dot.

---------

Co-authored-by: Ximon Eighteen <3304436+ximon18@users.noreply.github.com>