Rename `ServiceCommand` to `ServerCommand` and create `ServiceFeedback`.
This is needed because feedback comes from `Service` impls which don't know which Server passed them a request and will handle their response, and so they can only request reconfiguration of properties common to all servers., Commands however are sent to servers by the owner of the server and can reconfigure any aspect of the server, not just those common to all servers.

An example is DNS idle timeout reconfiguration which only applies to stream transports, not to datagram transports.

Also includes the following related changes:
- Fixes the wrongly swapped handling of `CloseConnection` and `Shutdown` in the stream `Connection`.
- Removes an irrelevant RFC related comment.