[base/scan] Implement 'Tokenizer::next()'
This hopefully comprehensive algorithm deals with quoted strings and
escaped characters therein as well as it can.  Of particular note is the
way it tries to insert newline tokens so that entries on different lines
can be distinguished.  It's difficult to do this for the last line of
the file, so we don't try.