Fix compilation bug with Rust 1.68.2 caused by an unnecessary HRTB (for<'a> ...) that Rust 1.69.0+ are fine with but 1.68.2 is not (perhaps due to https://github.com/rust-lang/rust/pull/103695/ which was fixed in Rust 1.69.0).
Also reorder XfrMiddlewareSvc generic types to match the order used in other middleware services, i.e. add the extra type XDP at the end, not before RequestMeta.

Also rename Metadata to RequestMeta as in other middleware services.

(and fix a comment that is ahead of its time, there is no ZoneMaintainer yet in this branch)