[new_name] Implement 'NameBuilder'
Unlike the existing 'NameBuilder', this type uses a fixed-size buffer
to write the name into.  This results in simpler code and it should be
more efficient.  It provides simple methods to extract domain names by
borrowing from the buffer instead of allocating.

This is a rewrite of <https://github.com/NLnetLabs/domain/pull/394>.