[new_base/parse] Rework '{Parse,Split}FromMessage'
Instead of passing the input selection as a range for parsing, the
whole message is cut (using 'Message::slice_to()') and only the start
is indicated.  This ensures that we never cross the end of the range.
It also implicitly dictates that compressed names are not allowed to
reference future locations in messages.

In addition, both the parsing traits now use offsets into the message
_contents_ rather than the whole message.  They can avoid 'as_bytes()'
everywhere and have better guarantees of success.  It also ensures the
message header can never be selected for parsing.