[new_net/server] Set up a service layering architecture
'Service' and 'ServiceLayer' are the most important traits defined here
('{Consume,Produce,Transform}Message' are auxiliary traits providing
high-level message parsing and building, specific to DNS server code).
The vast majority of the code is complex blanket implementations which
make these traits effortless to use.  As some examples:

- (ServiceLayer, Service): Service
- (Vec<ServiceLayer>, Service): Service
- (Vec<ServiceLayer>, Arc<ServiceLayer>, Service): Service

This makes it incredibly easy to compose services together, without
the restrictions of nesting that 'net::server' imposes.  Each service
layer is also much simpler, since it does not have to be generic over
the next layer in the pipeline.

The provided tuple implementations make it easy to use these layers in
static contexts, and it allows the compiler to inline everything into a
single function.

The auxiliary traits are also critical to the design -- they allow
messages to be parsed and built in a single iteration, preventing the
repetitive parsing of e.g. compressed names.  There are still open
questions regarding their flexibility for certain use cases.

The next step is to implement services and service layers using these
traits, and to use the traits from a transport implementation.