[new_base] Overhaul the building API
The mid-level building API has been completely rewritten.  It is much
simpler now, but no longer supports in-place modification or resumption
for the last message item.

'BuildIntoMessage' was renamed to 'BuildInMessage', and its signature
was completely changed to remove 'Builder'.  It is now quite similar to
'SplitMessageBytes'.  A slightly more involved truncation strategy is
used, where the name compressor is left to reset by the caller.