Add KMIP batching support.
Modifies the signing APIs to allow preparation of the bytes to sign
to be done separately to incorporating the generated signature bytes
into R1RSIG records.

Without this there is no way to batch KMIP signing requests as calls
to SignRaw::sign_raw() cannot continue without the resulting signature
bytes which will not be available until later when the rest of the batch
has been collected and submitted for signing.