Extend the serve-zone.rs example to make the mutations it performs more useful.
Prior to this change, the zone being served is mutated periodically but the SOA REFRESH duration is very high and no DNS NOTIFY messages are sent. This means that any nameservers acting as secondary to this zone will not see the mutations for a very long time. This change lowers the SOA REFRESH to cause secondaries to refresh their copy of the zone more frequently and thus receive the mutations.