Remove use of deprecated `ring::contant_time module`. (#584)
This PR removes the use of constant time slice comparison from ring as that
has been deprecated and replaces it with a new dependency on the
constant_time_eq crate.

This crate is used by a number of popular crates, most notable the zip and
blake3 crate. It has last been updated seven months ago. There seems to be
plans to move towards a 1.0 release. Given that this topic is fairly tricky
to get right, using this crate seems to be a better choice than implementing
our own version.

In order to be able to use the latest release of the crate, the minimum
supported Rust version needs to be bumped to 1.85, which this PR also does.

In order to unblock CI, the PR also fixes a complaint the nightly compiler
has about having two return statements if both ring and openssl are enabled.