Keyset improvements (#551)
Store the algorithm and key tag of a key to be able to reject duplicate
key tags and accidental algorithm rolls.

Store whether a key is considered available for a key roll. Rolls with new
keys that are not available are rejected.

Add to alternative key rolls for KSK and ZSK key rolls.

Add an algorithm roll.

Add more operations on UnixTime.

Add more actions
---------

Co-authored-by: Ximon Eighteen <3304436+ximon18@users.noreply.github.com>