Network servers should not ignore `ServiceError`. (#390)
This PR addresses an omission in the net::server::dgram and
net::server::stream modules whereby they iterate over the stream of service
responses accepting only Some(Ok) results and cease iterating on anything
else. To stop on None is fine, but to stop on Err without taking any action
is not.

Background: The Service interface was designed such that services can easily
return Err(ServiceError) with the inteion that this allows them to easily
produce certain standard DNS error responses without doing the work of
constructing the response message correclty, keeping their error handling
code paths simple. The ServiceError type has an rcode() method which the
network server was supposed to use to construct the appopriate DNS response
message, e.g. ServiceError::InternalError translates to rcode SERVFAIL.

But since the network servers never handle the service error case other than
to ignore it that means in such cases the client wouldn't receive any DNS
response message at all! This was not the intention, it was simply
overlooked and clearly there also aren't any test cases that exercise this.

This PR makes three improvements:

1.  Detec1 ServiceError and construct and return the appropriate DNS
    response error message to the client.

2.  Do this once in common code, factor out such common code from the
    network servers to a new ServiceInvoker base trait with default fn
    impls, with some network server specific methods to be filled in by the
    network server impls.

3.  Make the network server impls behave the same way for code that should
    already have been the same but through duplication had slight
    differences. E.g. till now the UDP server would spawn a background async
    task before even turning the message bytes into a Message object, while
    the TCP server would defer async task creation till a bit later.

This PR does NOT (yet?) address the lack of test cases for this logic.